Moving VMLA VM module to iree/modules/vmla/.
