+++ iree/hal/dylib/CMakeLists.txt
@@ -12,8 +12,4 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-if(NOT ${IREE_HAL_DRIVER_DYLIB})
-  return()
-endif()
-
 iree_add_all_subdirs()
