+++ iree/modules/vmla/op_kernels.h
@@ -31,8 +31,8 @@
 // semantics as reference and platform-specific versions can be implemented
 // as needed.
 
-#ifndef IREE_HAL_VMLA_OP_KERNELS_H_
-#define IREE_HAL_VMLA_OP_KERNELS_H_
+#ifndef IREE_MODULES_VMLA_OP_KERNELS_H_
+#define IREE_MODULES_VMLA_OP_KERNELS_H_
 
 #include <cstdint>
 
@@ -496,9 +496,9 @@ struct PoolingMax {
 
 // Inconsistent automated formatting here. Just disable clang-format (for now?).
 // clang-format off
-#include "iree/hal/vmla/op_kernels_generic.h"  // IWYU pragma: export
-#include "iree/hal/vmla/op_kernels_ruy.h"  // IWYU pragma: export
-#include "iree/hal/vmla/op_kernels_fft.h"  // IWYU pragma: export
+#include "iree/modules/vmla/op_kernels_generic.h"  // IWYU pragma: export
+#include "iree/modules/vmla/op_kernels_ruy.h"  // IWYU pragma: export
+#include "iree/modules/vmla/op_kernels_fft.h"  // IWYU pragma: export
 // clang-format on
 
 #endif  // IREE_HAL_VMLA_OP_KERNELS_H_
