+++ iree/modules/vmla/op_kernels_ruy.h
@@ -12,8 +12,8 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#ifndef IREE_HAL_VMLA_OP_KERNELS_RUY_H_
-#define IREE_HAL_VMLA_OP_KERNELS_RUY_H_
+#ifndef IREE_MODULES_VMLA_OP_KERNELS_RUY_H_
+#define IREE_MODULES_VMLA_OP_KERNELS_RUY_H_
 
 #include <type_traits>
 
@@ -121,4 +121,4 @@ Status MatMul::Execute(RuntimeState* runtime_state,
 }  // namespace hal
 }  // namespace iree
 
-#endif  // IREE_HAL_VMLA_OP_KERNELS_RUY_H_
+#endif  // IREE_MODULES_VMLA_OP_KERNELS_RUY_H_
